crash: Including /var/log/cros_ec.log for EC info logging.
The upstart job "crash-boot-collect" is now initiated after the
"ecloggers" job has stopped. This is due to the EC and kernel collectors
in the crash reporter relying on the timberslide service to gather and
store EC logs, and the timberslide service is triggered by the
"ecloggers" upstart job.

BUG=b:378671179
TEST=unittest

Change-Id: I7b362d7aed449500ae2c554134ff85f494da160e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6313999
Tested-by: Troy Wang <troywang@google.com>
Reviewed-by: Jong Ahn <jongahn@chromium.org>
Commit-Queue: Troy Wang <troywang@google.com>