power: Fix missing dbus response sending
Currently, the generated dbus code makes method call messages
without `dbus_message_set_no_reply`. This means that dbus-daemon
expects these method call messages will get a reply from parties
being called. If replies are not sent, the `pending_reply` will
not be cleared in the dbus-daemon. There is a per DBusConnection
limits on the number of pending_replies. When the limit is hits,
the DBusConnection will not be able to receive any method calls.

This CL adds the missing response sending in the found places
in `powerd`.

BUG=b:399125955
TEST=Manual. Lower limits in system.conf say 10, login and wait.

Change-Id: I6b2b05803d06059fdc07240121f2d9bb7acb8815
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6318395
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>
Tested-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>