installer: Do not remove ureadahead pack after update
Ureadahead acquired a new feature called `record-and-replay`
recently, which continuously updates the pack file while
also utilizing the old pack file, and replacing the stale pack
by newly generated up-to-date pack in the end.

This effectively keeps the pack file always up-to-date, which
eliminates the necessity of removing a pack file to force the
ureadahead to regenerate a pack.

BUG=b:380800396
TEST=CQ
TEST=`build-image` and `cros flash`
TEST=`cros flash --no-reboot`, and examine `/var/lib/ureadahead` directory.

Cq-Depend: 6381873
Change-Id: I733385111cd0631770c77353436351a143b20489
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6382112
Tested-by: Soichiro Fujii <soichirof@google.com>
Commit-Queue: Soichiro Fujii <soichirof@google.com>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>