shill: network: always do gateway pings in ConnectionDiagnostics
This patch changes the logic to trigger a ping to the gateway in
ConnectionDiagnostics. Instead of doing conditional ping requests to the
gateway (for example, after a failure to ping the target server), that
ping request is always initiated at the beginning of
ConnectionDiagnostics.

The results of DNS server ping requests are also ignored for selecting
with servers to use for resolving the name of the target server.

BUG=b:229309479
TEST=FEATURES=test emerge-$BOARD shill; CQ

Change-Id: I01e30877bc0abce4625eb9c01ab8fe74d35f0660
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6335434
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>