crash: Include console-ramoops when kcrash is corrupted
dmesg-ramoops is compressed and thus when it's corrupted, it can be hard
to recover it. This patch adds the logic to include console-ramoops in
this case. Since console-ramoops is uncompressed, it might provide more
information on why the corruption happened and how bad the corruption
is.

In the corruption case:
- Before this change, the corrupted dmesg-ramoops will be collected as
  the .kcrash file;
- After this change, the console-ramoops will be collected as the
  .kcrash file, and the corrupted dmesg-ramoops will be collected as the
  .corrupted file.

BUG=b:339509630
TEST=Manually verified that when the kernel left the enc.z file, the
  collected content under /var/spool/crash is expected.

Change-Id: Ife28b4ab6a90c81fa6db6fc592a376b83ac12f25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6310962
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>