ml_core: cacher: Allow clock_gettime seccomp policy
A seccomp violation is observed in the report of b/400516876 due to the call of clock_gettime.
Besides, we also have a similar bug happen on the x86 platform (b/362365444).
Therefore, it is clear that we need to allow clock_gettime in our seccomp policy.

BUG=b:400516876
TEST=CQ

Change-Id: I2ea546c95fe025d05b12d31b550da5c6c104de35
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6334221
Tested-by: Tommy Chiang <ototot@google.com>
Reviewed-by: Yu-Che Cheng <giver@chromium.org>
Reviewed-by: Imran Ziad <imranziad@chromium.org>
Commit-Queue: Tommy Chiang <ototot@google.com>
Auto-Submit: Tommy Chiang <ototot@google.com>