cros-codecs: Add token to GenericDmaVideoFrame
Swap drop callback model of FFI support for GenericDmaVideoFrame in
favor of an RAII model where we directly include the FFI'd object in the
GenericDmaVideoFrame. The advantage is that we are able to arbitrarily
access the FFI object throughout the lifetime of the
GenericDmaVideoFrame instead of simply destroying it.

BUG=b:397758611
TEST=atest ccdec_test

Change-Id: I9a0e4b07d58c48e6494983d2b4ade307645f6d28
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6599575
Tested-by: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Auto-Submit: Bijan Choobineh‎ <bchoobineh@chromium.org>
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>