vm_tools: fix stateful_device for Baguette VMs
The stateful disk (which is also the rootfs) on containerless VMs is
/dev/vdb, but the TerminaVm startup code was setting it to /dev/vdc,
which does not normally exist for containerless VMs. Additionally,
containerless VMs don't run the normal maitred initialization code, so
re-add the fallback to /dev/vdb in maitred's ResizeFilesystem
implementation.

This is a bit of a hacky workaround rather than a proper fix, but it is
the shortest path to working disk resize for containerless VMs.

BUG=b:401551934
TEST=enable #crostini-containerless and resize VM disk via UI

Change-Id: Id730fd418e93c610194d945337c0782dae672e82
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6336806
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>