init: Add preservation helper for test/dev flows
Add a helper that can preserve preseeded file metadata in the
metadata partition. The helper is intended to be called
in test flows that may result in a reset of the system encryption
key.

This helper is invoked during the (dev) shutdown path to dump the
state of preseeded files that should be preserved. On subsequent
boots, if the TPM state is reset, the default-key-stateful
filesystem will be reset and the files from the last boot will be
preseeded. If the TPM state is not reset, the preseeding logic
will delete the stored state.

Additionally, do not attempt to reinitialize existing files. This
requires a slight fix in the existing logic which was mistakenly
first creating a directory instead of the file.

Fixes: fe745fc7356f6 ("init: Add FilePreseeder")

BUG=b:399570398
TEST=gbb transition from 0x0 <-> 0x9 with preserve_dev_image
TEST=tast run hwsec.FWMPAcrossTPMClear passes

Disallow-Recycled-Builds: test-failures
Change-Id: Ib828af71dcf9be4ff14ec2d6e01c1de75fa3775f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6340228
Commit-Queue: Sarthak Kukreti <sarthakkukreti@google.com>
Tested-by: Sarthak Kukreti <sarthakkukreti@google.com>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Auto-Submit: Sarthak Kukreti <sarthakkukreti@google.com>