[owners] Remove vincentchungzt@google.com from uwbd/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/vincentchungzt@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I848953bc55486763214a936120e83a49cc02734b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6343795
Commit-Queue: Ningyuan Wang <ningyuan@google.com>
Auto-Submit: swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Tested-by: Ningyuan Wang <ningyuan@google.com>
Reviewed-by: Ningyuan Wang <ningyuan@google.com>