patchpanel: Update redirect_user_dns jump rule
Previously, the jump rule to the chain is statically added through
patchpanel's iptables template. This patch updates the rule such that it
is dynamically updated with the lifetime equals to the dns-proxy system
proxy's lifetime.

redirect_user_dns chain is used to DNS traffic to be redirected to the
default proxy of dns-proxy. However, the behavior of default proxy
running without the system proxy running is undefined.

This patch updates the behavior such that the default proxy is only
operational (i.e. the redirection is enabled) when the system proxy is
running. By doing so, the system will fully rely on the default behavior
(no DNS proxy), avoiding the undefined behavior.

One possible trigger of this case is VPNs, see the bug for details.

BUG=b:401457445
TEST=manual;
     observe that after removing all the redirection rule of system
     proxy and updating /etc/resolv.conf, connectivity is still
     available with VPNs.

Change-Id: If459ff1cba527199ee85f0d31dc728234d5df11b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6346486
Commit-Queue: Jason Iman <jasongustaman@chromium.org>
Tested-by: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>