patchpanel: Set dst when applying a blackhole route
Currently, when patchpanel installs a blackhole route with a non-default
prefix, it will only set the prefix length field in the RTNL message,
but leave the address part empty, which results in a wrong route in the
routing table. This is a regression introduced by the patches for
b/328153824 and b/368125566.

This patch fixes this issue.

BUG=b:402165537
TEST=tast run $DUT network.ARCVPNDatapath network.ARCVPNSplitRouting*
TEST=manually tested with the Zscaler app.

Change-Id: I6d5a9a76258cbc77dca76dff84291b79bea338c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6345352
Reviewed-by: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>