cryptohome: Disable pinweaver passwords in CreateVaultKeyset
There's no reason you'd ever encounter a pinweaver password vault keyset
in the wild, as versions that support pinweaver passwords and versions
that create vault keysets are mutually exclusive.

This requires special-casing the VK creation code to exclude this
feature, but there isn't really a better way to do this as none of the
generic creation code has any understanding of being in the context of a
VK creation, and it wouldn't really make sense to add that kind of
context either since this is really only an API for compatibility
testing.

BUG=b:404549110
TEST=unit tests, tast tests with pinweaver passwords enabled

Disallow-Recycled-Builds: all
Change-Id: Ia38a2421668a1de9d17129bd8b12101f54d4a13f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6369077
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>