cryptohome: Make unit tests more pinweaver password clean
This fixes up a bunch of unit tests to make them more agnostic as to
whether or not pinweaver passwords are enabled. For the most part this
is just injecting default IsPinWeaverEnabled behavior, keeping it off
because most of these tests are explicitly testing non-pinweaver
implementations.

There are a few tests which still fail if pinweaver passwords are
enabled. However, I don't think it makes sense to "fix" them, they're
checking default behavior in a way such that I think the correct fix is
to update them when we change the default.

BUG=b:391427550
TEST=unit tests, also ran them with the feature default flipped

Change-Id: I4c9e4422a060151c887495bac22d3c4667ccf62f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6370077
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>