shill: wifi: Fix connection metrics
In CL https://crrev.com/c/5850668 calls to ServiceDisconnected() were
introduced in couple failure scenarios for two reasons:
- to make sure the service is properly marked as disconnected,
- and to generate FW dumps (as a side effect of this call).

This however introduced possibility of double calls of this function for
a particular service which leads to incorrect metrics collection.  So
instead the previous logic is kept and the firmware dumps are generated
in the problematic places directly.

BUG=b:402463786
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I1c2292ef4e43767f967871e96033fbf318bc2b16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6384395
Commit-Queue: Andrzej Ostruszka <andrzejo@google.com>
Tested-by: Andrzej Ostruszka <andrzejo@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>