[owners] Remove zhaon@google.com from ml_core/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/zhaon@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I277881ba0f0f8db8e0df8733eaa666cd0d1cc3bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6391006
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Nikolas Bowe <nbowe@google.com>
Tested-by: Nikolas Bowe <nbowe@google.com>
Reviewed-by: Nikolas Bowe <nbowe@google.com>