cros-codecs: Query for DecodedFormat
Prior to this CL, the DecodedFormat set was hard coded to MM21. While
this will work for some streams, it will not work for 10 bit streams for
example where a format of MT2T is needed.

Updates the stateless decoder logic to query the CAPTURE queue for the
DecodedFormat for the stream. Additionally, if the decoder detects that
the bitstream has a bit depth of 10 bits, it will send a fake header to
the OUTPUT queue to let the device know that the stream will be 10 bit.

BUG=b:401035899
TEST=Was able to run test-25fps.vp9 which has a bit depth of 8 bits on
Geralt and was able to see via logging that the DecodedFormat was MM21.
Also ran test-25fps.vp9_2 on Geralt which has a bit depth of 10 bits.
Verified through logging that the chosen DecodedFormat chosen was MT2T.

Change-Id: Ia1ffd78b7516d79cd4518428d2b1a82ef3d59e19
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6377270
Auto-Submit: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Bijan Choobineh‎ <bchoobineh@chromium.org>