cros-codecs: Update enumerate_devices() logic
Prior to this CL, the enumerate_devices() logic would only check that
there was a matching media device for the selected video device. This
works on some devices, however other devices have this logic accidentally
picking devices such as the camera since there is a matching media
device.

This CL updates the |enumerate_devices()| function to take in a Fourcc
value which it will query the OUTPUT queue of each device to see if it
is supported. Unfortunately, the Fourcc value is not available at the
time of creation of the DeviceHandle. As such, the reference in the
V4L2Device is updated to be an Option which will be created in the
initialize_queues() phase.

This CL results in being able to select the correct device on boards
such as Geralt and being able to decode bitstreams on those boards.

BUG=b:401035899
TEST=Ran test-25fps.vp9 through ccdec on a Geralt successfully.
Visually inspected logs to verify that enumerate_devices was picking the
correct device.

Change-Id: I7b7f5c1492a446530029b724456a1ca4242fd9e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6398042
Tested-by: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>
Auto-Submit: Bijan Choobineh‎ <bchoobineh@chromium.org>