patchpanel: ndproxy: ignore VLAN tagged ICMPv6
This patch uses the PACKET_AUXDATA socket option on the packet socket of
NDProxy to detect and filter out VLAN tagged packets.

See the Linux manual pages for packets(7) and cmsg(3) for more details
on how to obtain the struct tpacket_auxdata when using recvmsg and how
to read the auxiliary packet data.

BUG=b:400590479
TEST=FEATURES=test emerge-$BOARD patchpanel, CQ

Change-Id: If4664ba86fd869abc7c25e49463e35e2734f98a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6414931
Reviewed-by: Jason Iman <jasongustaman@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>