installer: Add new rust binary install-copy-esp
On devices with 4k blocks, the traditional bitwise copy of the ESP
doesn't work: it produces an ESP that can't be mounted.

Add a rust binary that can copy an ESP on 4k devices by:
* Creating a new vfat filesystem on the destination partition
* Copying over all the files from the source ESP

One thing this doesn't handle: BIOS/'legacy boot', which depends on
a "VBR" (Virtual Boot Record) at the start of the partition. Simply
copying the VBR over with a new filesystem doesn't seem to work,
syslinux (which we use to create the VBR) includes data in there that's
unique to the filesystem. To update this we'd probably need to include
syslinux in base images and call it on the new ESP.

A later commit will actually use this binary.

BUG=b:399457884
TEST=cargo clippy && cargo build && cargo test

Change-Id: I1d4d3eb2dd03462ff93374a8d2736553ade3de4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6441223
Commit-Queue: Ted Brandston <tbrandston@google.com>
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Ted Brandston <tbrandston@google.com>