crash: Save OS ver in kernel log and prioritize its use in kn-collector
The kernel collector retrieves the kernel log (ramoops dmesg) as the
kernel dump. This CL ensures that the current OS version is logged to
the kernel log immediately after boot while OS version is saved to file.
This helps the kernel collector accurately determine the OS version at
the time of a crash.

BUG=b:346426906
TEST=`cros_sdk emerge-brya chromeos-base/crash-reporter`
TEST=`cros_workon_make --test --board=brya chromeos-base/crash-reporter`

Change-Id: I807141d97a62736873ae80e4e5ab9a083a63f7bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6419439
Commit-Queue: Troy Wang <troywang@google.com>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Tested-by: Troy Wang <troywang@google.com>