shill: cellular: Update the default IP_TYPE to IPV4V6
Previously, when an IP type wasn't explicitly defined in the profile,
we defaulted to IPv4. However, given the growing prevalence of
dual-stack and IPv6-only networks, we are updating our
default to IPv4v6 (dual-stack) in such cases. This ensures better
compatibility with modern network infrastructure.
It's important to note that for roaming scenarios,
we will maintain the IPv4 default, as this is the behavior on Pixel devices.

BUG=b:358751309
TEST=CQ

Change-Id: I48bd3a9d5cc75534f6b757e991709802ac9bb0ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6427482
Tested-by: Madhav <madhavadas@google.com>
Commit-Queue: Madhav <madhavadas@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>