net-base: validate IPv6 RA DNSSL domains
This patch ensures that the domain search list found in the IPv6 RA
DNSSL option are valid ascii strings, implying there are also valid
UTF8 strings and can be serialized in DBus messages.

This patch also ensures that the domain search list is not printed in
clear in logs because it can constitute PII data.

BUG=b:408883419
TEST=FEATURES=test emerge-$BOARD net-base. CQ

Change-Id: I7cbdcdebd03d780ba3ddec23bffa906b09d7e708
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6440522
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>