cryptohome: Restore some of the backend cert tests
This gets rid of the tests that are date-dependent, but restores the
rest. The initial change that this is partially rolling back was done as
a quick-fix, this is a little less drastic.

I took a look at trying to make the test more hermetic by adjusting the
time that is used as current time for the verification, so that we could
make the verification happen as if it was always running on the same day
(a day where the certs are still valid) but this would require a
non-trivial refactoring of all the code which doesn't seem justified
under the current state of maintenance.

BUG=b:409607120
TEST=unit tests

Change-Id: Ief7f3315cadf1746f1a4e0d8b35ab14ecd93fd3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6444323
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Auto-Submit: John Admanski <jadmanski@chromium.org>