installer: Wipe metadata partition when not in use on update
Since it is possible to rollback to an image that do not support
dm-default-key, updating to an image that supports it will trigger
a clobber.
Prevent the issue by wiping the meta-data partition on update when
stateful is mounted on top of a dm-default-key volume.

BUG=b:409836532
TEST=CQ. On a nissa DUT targeted for dm-default-key with latest (R135),
check dm-default is used.
cros flash  latest R134 (R134-16182.8.2)
Verify dm-default-key is not used but metada partition is valid.
Check clobber happened (sda1 unreadable, appears corrupted).

Install new image. Check partition 11 is wiped. Reformated with
a new UUID.  Check clobber occurs back to dm-default-key.

Change-Id: I530cd6053bc71df89b73f1febf3f06814455accd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6451959
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>