crash: Check additional release version in kernel dump to decide OS ver
This CL is an additional fix to
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6419439

There is a hole in the previous implementation. When the device is first
updated from an OS version without this new logic, and it happened to
crash on the old OS after crash collector runs, but before the log
overflows, crash_reporter will report the new OS version, which is
incorrect.

BUG=b:346426906
TEST=`cros_sdk emerge-brya chromeos-base/crash-reporter`
TEST=`cros_workon_make --test --board=brya chromeos-base/crash-reporter`
TEST=Manually tested.

Disallow-Looks-For-Green: True
Change-Id: I73e837fe50c3180b153b029e71c0dc43eaa0e750
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6452212
Commit-Queue: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Tomasz Figa <tfiga@chromium.org>