flex_hwis: Add upstart service to copy boot log from UEFI variable
Crdyboot, the Flex bootloader, writes a verbose log to a UEFI
variable. Copy this into a regular log file under /var/log for ease of
access.

This service is not directly related to HWIS, but is similar to the
flex_device_metrics previously added to this directory.

Note that adding this file to `/var/log` does not automatically add it
to feedback reports.

BUG=b:405968720
TEST=cros_workon start chromeos-base/flex_hwis
TEST=emerge-reven chromeos-base/flex_hwis
TEST=Deploy chromeos-base/flex_hwis to reven image
TEST=Boot with UEFI, verify expected log file is created
TEST=Boot without UEFI, verify service exits successfully

Change-Id: I7653caef019b167114220458006f5a303c46cdd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6453699
Tested-by: Justin Ouellette <justinoue@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Justin Ouellette <justinoue@google.com>