shill: DHCP: Collect provision reason and emit result to UMA
Collects DHCP provision reason in DHCPController and emits the result
to UMA. To collect the reason, we adds a reason argument in the
following functions so that the reason can be passed smoothly:
DHCPController::RenewIP(), Network::RenewDHCPLease(),
Device::ForceIPConfigUpdate().

Note that we also add a kConnect reason to
dhcp_pd_controller_->RenewIP(). The reason is DHCPv6 also uses
DHCPController and there has to be a reason in RenewIP().
We don't emit metrics for DHCPv6.

For Device::ForceIPConfigUpdate(), adding a reason argument is actually
not necessary because this function is only used in suspend/resume, but
the function name itself does not imply its actual usage, so we also
passed a reason to it.

We store the collected reason as the provision_reason_ property in
DHCPController. We reset this property when emitting metrics so
that at most one metric can be sent per provisioning event.

The difference between emitting kNak and kTimeout lies in whether any
NAK is received during provision. We use a property nak_received_ to
record whether any NAK has been received. We create helper functions
UpdateProvisionStatus() and ResetProvisionStatus() to help update and
reset provision_reason_ and nak_received_ together.

BUG=b:397523142
TEST=unit

Change-Id: Ic4e0627f80297b3533855aadeeef41d0ef6fe45e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6481673
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Zikai Chen <chenzikai@google.com>