avtest_label_detect: Detect protected content capabilities only if ENABLE_L1HWDRM
We detect protected content capabilities on MediaTek and AMD if
the Mediatek secure heap node exists and AMD-TEE exists, respectively.
These are available by default on newer kernels.
We don't enable protected content playback on old MTK and AMD boards.
As the protected content capabilities are used to run the tast tests,
we should not detect the capabilities even if the kernel is feasible
for L1 HWDRM.

BUG=b:412340224
TEST=video.Capability on zork and guybrush

Change-Id: I14455e01d5cc5b818964575654a02b44f99fb121
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6492250
Tested-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>