cros-codecs: Store DMA buffer mapping lens for munmap
Previously, we were calling munmap() using the length of the plane as
the argument, but we should be calling munmap() with the length of the
entire mapping to avoid leaking memory. In contiguously allocated video
frames, the length of the entire mapping is usually equal to the size of
the entire allocation, so it frequently does not match the length of the
frame.

BUG=b:413646151
TEST=compiled on Brya

Change-Id: Icf9ea5f45c55f7e8520089bb04cf29536ba678d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6599625
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>
Auto-Submit: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Tested-by: Bijan Choobineh‎ <bchoobineh@chromium.org>