installer: Fix BLKRRPART EBUSY error
In `reload_partitions`, `blockdev --rereadpt` can fail after `udevadm
settle` with an EBUSY error. If this occurs, sleep for a short time and
then retry. Only a single retry is used since testing shows this is
sufficient, and to avoid any unnecessary complexity in a shell script.

Testing: to quickly test this error, insert an `exit 0` after the call
to `reload_partitions`. Run the installer in a loop 100 times. Without
this fix, the EBUSY error will occur at least once, usually within a
dozen iterations or so. With this fix, reload always succeeds after the
retry.

BUG=b:410051948
TEST=See above; tested on the x1cg6

Change-Id: I2ee37fbc170fc75b3d54d7aa88d44a71cf0195bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6502952
Reviewed-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>