[owners] Remove yich@google.com from device_management/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/yich@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I5a5a8078584bff05e179343705f105fcf34652f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6505906
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Howard Yang <hcyang@google.com>