cros-codecs: Create v4l2 capability logic
This CL introduces new capabilities logic which will be used throughout
cros_codecs and codec2 to do unified checking of the decoders
capabilities.

This change creates a |DecoderCapability| struct which will hold the
decoders capabilities for a given codec. This change also introduces
logic that can be used to query all the supported capabilities for the
decoder and implements it for v4l2 devices.

BUG=b:411419404
TEST=N/A

Change-Id: I46bb3783da11cbe70f9c0b7897020b57aefdd89e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6599645
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>
Tested-by: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Auto-Submit: Bijan Choobineh‎ <bchoobineh@chromium.org>