libbrillo: Handle configless field in DecodeHWID
Modify DecodeHWID to correctly parse and decode HWIDs that include an
optional configless field.

Also, updates the test cases of DecodeHWID to make the checksum valid.
This does not change the behavior of the test, but is intended to make
the test more consistent with the name of the test case
"DecodeValidHWIDSuccess"

BUG=b:367899338
TEST=FEATURES=test emerge-${BOARD} libbrillo

Change-Id: I69ed7cd5e4abeed6c040c3e0efc143c30a02f03a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6508008
Tested-by: Allen Shih <allenshihmc@chromium.org>
Commit-Queue: Allen Shih <allenshihmc@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>