cros-codecs: Temporarily use GenericDmaVideoFrames for VA-API auxiliary pool
The upgrade to kernel 6.12 and the XE driver seems to have broken
gbm_bo_map() on VA-API machines. Because we rely on this function for
to perform NV12->I420 conversions in our ccdec_tests, this was causing
failures. As a temporary workaround, we can export the GbmVideoFrames to
GenericDmaVideoFrames.

BUG=b:416258113
TEST=atest ccdec_vp8_test on Brya

Change-Id: Ia13b58f962717fef2a21b871f648062ed04d8d5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6599642
Tested-by: Bijan Choobineh‎ <bchoobineh@chromium.org>
Auto-Submit: Bijan Choobineh‎ <bchoobineh@chromium.org>
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>