installer: Start centralizing postinstall env var reads
The post install type is read from several env vars (or, in the case of
updates, implied from the lack of env vars and not running
minios). Rather than reading directly from the environment, put these
values in InstallConfig. This will make testing easier, since
InstallConfig is already passed around everywhere.

The env vars are now read in ConfigureInstall, which is the central
location where the InstallConfig is constructed in non-test code.

It seems likely these bools could be consolidated into a single enum
value. Left a TODO to that effect, but for now keeping the separate
bools to avoid too much churn.

For now, only ChromeosChrootPostinst is updated to use the new
InstallConfig fields; other code will be updated in later commits.

BUG=b:416310277
TEST=FEATURES=test emerge-reven chromeos-installer

Change-Id: I412ec3d5027776d4978843bb2f58d2a0160cac90
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6547203
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Andre Braga <andrebraga@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>