installer: Add Platform abstraction to postinstall
Similar to other code in platform2, the Platform class is an abstract
interface, and MockPlatform can be used in tests to mock out platform
functionality.

This class will be used to enable more unit testing of postinstall
code. Initially it just provides DumpKernelConfig, with the same
interface as the function in inst_util.h.

BUG=b:416310277
TEST=FEATURES=test emerge-reven chromeos-installer

Change-Id: I042d91501e93fcccb1f7e46aa7b2d3da9c1664f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6553464
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>