cryptohome: Fix exposing recovery seed
A recovery seed is a unique identifier used to generate recovery IDs.
Previously, the last recovery ID was incorrectly shown in the UI instead
of the recovery seed. This CL corrects this by ensuring the recovery
seed is shown and includes new tests to verify this behavior.

BUG=b:378357889
TEST=CQ

Change-Id: Icbc35e6e7399371d87279d6a1124ebb71efd0c83
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6556809
Commit-Queue: Mohammed Abdon <mohammedabdon@chromium.org>
Tested-by: Mohammed Abdon <mohammedabdon@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>