dlp: Run unit test as not parallel.
Recent change in dlp threading made the test much more flaky
when run in parallel (when multiple DlpAdaptors are created).
As the flakiness is reproducible only in trybots, changing
the test to run in a single job to confirm the suspect.
The unit tests run really quickly, so there should be no
delay due to this change.

BUG=b:419436152
TEST=cros_run_unit_tests --packages chromeos-base/dlp

Change-Id: I386565b8fa0eded3e6d6aa8bd5b6159b0a3862f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6578262
Reviewed-by: Brett Brotherton <bbrotherton@google.com>
Auto-Submit: Sergey Poromov <poromov@chromium.org>
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Tested-by: Sergey Poromov <poromov@chromium.org>