cros-codecs: c2_encoder: Copy the H.264 SPS/PPS as CSD
This CL adds code to copy the SPS/PPS NALUs from the encoded bitstream
for the first completed C2EncodeJob and then store their concatenation
separately from the complete encoded bitstream. This is so that the C++
HAL can copy that into a C2StreamInitDataInfo::output which serves as
initialization data for the client and is needed for VTS tests.

BUG=b:389993558
TEST=VtsHalMediaC2V1_0TargetVideoEncTest on brya

Change-Id: If990a437760f4ff87753d3421bb617067e7ebb97
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6674646
Auto-Submit: Bijan Choobineh‎ <bchoobineh@chromium.org>
Tested-by: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@google.com>
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>