[owners] Remove imranziad@chromium.org from camera/diagnostics/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/imranziad@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Iecc08790df4c820ddfa734dbc9b68138e4939a3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6591396
Reviewed-by: Shik Chen <shik@chromium.org>
Commit-Queue: Shik Chen <shik@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Tested-by: Shik Chen <shik@chromium.org>