[owners] Remove imranziad@google.com from ml_core/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/imranziad@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Iec7886fda6f9c9f25921a5b82b58c94306960f06
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6591595
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Tommy Chiang <ototot@google.com>
Commit-Queue: Tommy Chiang <ototot@google.com>
Tested-by: Tommy Chiang <ototot@google.com>