dlcservice: Prevent truncation of pending DLC images
.. during OS / repeated OS updates.

This change prevents the corruption of pending DLC images when an OS
update is in progress and the new DLC image is larger than the one from
the current build.

When the OS is scheduled to update from build A to build B, a user
might also update a DLC. If the DLC image for build B is larger than for
build A, a problem occurs. The dlcservice on build A, when preparing the
inactive slot for the update, would create the new image file but
truncate it to the smaller size specified in build A's manifest.

This truncation would chop off the final bytes of the larger DLC image
being downloaded for build B. As a result, the dm-verity hash tree
appended to the end of the image becomes corrupted.

After the OS update to build B completes, this corruption is detected,
forcing a full re-download and reinstallation of the DLC.

To fix this, a no_truncate flag is introduced in the CreateFile utility.
When preparing the inactive DLC image, this flag is enabled.
This ensures that if a file already exists and is larger than the size
known to the current build, it will not be shrunk, thereby preserving
the integrity of the fully downloaded image for the new build.

BUG=b:418563926
TEST=unittest + builders/CQ

Change-Id: I4dab12e882d4e47d75cc155db7ab44e9f53daf86
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6604748
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: TingHao Lin <tinghaolin@google.com>
Auto-Submit: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: TingHao Lin <tinghaolin@google.com>