cros-codecs: Relax general_level_idc restrictions for h.265
The ITU specifically state that general_level_idc is not valid if
the value parsed is not in a precribed list. However, clips exist
that do not meet this standard.

Instead of rejecting the clip outright, allow for a value of 0
and interpret it as level 1.

BUG=b:419500601
TEST=run previously failing clips

Change-Id: I45ac18793e2041e88c44936519cbef70a0c01554
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6674639
Auto-Submit: Bijan Choobineh‎ <bchoobineh@chromium.org>
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Hang Nguyen <hnt@google.com>
Tested-by: Bijan Choobineh‎ <bchoobineh@chromium.org>