[owners] Remove jkardatzke@google.com from hwsec-optee-ta/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/jkardatzke@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I151948a66e08fa943da4b0a219558741e2586c80
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6617841
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Yi-An Chen <chenyian@google.com>
Reviewed-by: Yi-An Chen <chenyian@google.com>
Tested-by: Yi-An Chen <chenyian@google.com>