libarc-attestation: Force new certificate for ARC
This CL modifies the existing functions for
provisioning ARC certificates.

If the certificate does not show a verified state and
issue date of the certificate is before a cut-off date,
a new forced certificate will be requested.

Overall, this will help with the recovery of incorrect
certificates that may have been provisioned to a lot of
ChromeOS devices as the device state has been
included in the certificate at the cut-off date.

BUG=b:418122776
TEST=FEATURES=TEST emerge-hatch libarc-attestation

Change-Id: Ia68253463b08c94c97aecdaa87453a8ae5fcd881
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6622438
Reviewed-by: John L Chen <zuan@chromium.org>
Commit-Queue: Vaibhav Raheja <vraheja@chromium.org>
Tested-by: Vaibhav Raheja <vraheja@chromium.org>